Login feature
Make fxml file. Works bu check the ID and password, if ID is wrong it won't check the password. Also make it possible to jump from login-register page. Add temporary main page to see if login works or no.
There is a room to continue working on which is make it possible to return user object to identify who is the user that is currently using.