The main page with navigation bar
I make this by making a knock off version of FXRouter. Basically, it works by instead of setting the Stage which is the entire window to be the main page, I set the Stackpane in side the main page to be the main page that will be resposible for changing the window instead. The program should work this way imo. My teammates won't have to change anything much. They just have to import the FXRouterPane instead of the normal one. Everything should still be working normally.