Admin Page
I made the fxml and the controller. The main function of this page is to look at the log in history of the user.
On the right, there is a summary info of that user. (Not developed yet) If you click at the View Profile button,
I will lead to more information of this user such as login info, events information.