UPLOAD PROFILE PIC + EVENTLIST
Due to some reason, I could not finish the upload profile pic and go to event list, so this commit will contain with upload pic and EVENTLIST.
Eventlist work by fetch the event list from csv file then use the data to create an fxml file and the load that to the gridPane in EventListController.
EventElement exists for helping to create a children to put in gridPane.